[Feature #157262513]run migrations to create database tables according to models
